feat(attendance): ADR-0016 Phase 4 — fusion modèles/services/commandes SmartAttendance (Closes #5355)
