+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - **feat(attendance/i18n+tests+docs): pointage — messages API localisés ×4, tests rapports par période, guide utilisateur (Closes #5269).** 5 littéraux de messages (workflow d'approbation, calendrier, événements géo) → catalogue `attendance.*` (×4 fr/en/tr/ar, valeurs EN identiques aux littéraux historiques ; codes machine API exclus) ; `AttendancePeriodReportsTest` (rapports day/week + filtre employé + export CSV + RBAC, moteur #5268) ; `AttendanceI18nTest` (parité ×4 + contenu) ; guide utilisateur `docs/user-guide/POINTAGE.md`.
 
 - **perf(accounting): audit perf/scale — index, barrière N+1, benchmark 10k documents (Closes #5275).** Migration additive 000005 : index composés `accounting_documents (company_id, status, due_date)`, `(company_id, issue_date)` et `accounting_payments (company_id, document_id, status)`. Commande `artisan accounting:benchmark {--documents=10000}` (protocole F-12) : seed 10k documents réalistes (entreprise dédiée, inserts groupés) + mesures — liste paginée eager avec **barrière N+1 ≤ 5 requêtes** (indépendante de N), requête relances J+7, agrégation mensuelle. Résultats réels : **liste 10k docs en 0.01 s** (cible < 200 ms ✅), relances 1 763 éligibles, 13 mois couverts. `AccountingPerformanceTest` (3 tests : N+1, chemin indexé relances, agrégation) + `docs/accounting/BENCHMARK.md` (protocole, cibles, index, tableau de référence).
+- **feat(attendance): ADR-0016 Phase 4 — fusion modèles/services/commandes SmartAttendance dans le module Attendance (Closes #5355).** 4 modèles (`GeoAttendanceSession`, `EmployeeLocationEvent`, `AttendanceModeSettings`, `EmployeeAttendancePreference`), 2 services (`AttendanceModeResolver`, `GeoSessionManager`), 6 actions, 2 DTOs, 3 exceptions et le contrat `GeofenceValidatorInterface` déplacés dans `App\Modules\Attendance\*` avec classes de re-export `@deprecated` dans SmartAttendance (imports existants préservés). `AutoCloseGeoSessionsCommand` fusionnée dans `AutoCloseAttendanceCommand` (une seule fermeture automatique — pointages sans check-out + sessions GPS orphelines, options `--hours`/`--company`/`--logs-only`/`--sessions-only`) ; `smart-attendance:auto-close` conservée comme alias délégué `@deprecated` ; scheduler unifié sur `attendance:auto-close`. Tests : 6 fichiers `tests/Feature/SmartAttendance/*` migrés vers `tests/Feature/Attendance/Geo*` sans perte de scénarios. Garde géofence (`check-geofence-single-usage.sh`) mise à jour. Aucune route `/smart-attendance/*` cassée (surface conservée jusqu'à la Phase 5 #5356).
 - **feat(mobile): dé-duplication lot 2 — grappe user_auth complète (repository + provider + 3 écrans) + team/tasks providers extraits vers leopardo_core (progrès #5279, suite #2601).** 7 paires supplémentaires éliminées (14 instances) : `features/user_auth/{data,providers,screens}` (repository unifié — les versions hr/manager ne différaient que par le formatage) + `team/providers/team_provider.dart` + `tasks/providers/task_provider.dart`. `userAuthRepositoryProvider` déplacé dans `leopardo_core/core/providers/core_providers.dart` (partie commune). Dépendances ajoutées au core : `google_sign_in ^7.2.0`, `flutter_animate ^4.5.2`. `flutter analyze` 0 issue ×3 packages ; tests 100/100 (hr 35, manager 34, employee 31). Cumul lots 1+2 : 17 paires éliminées (34 instances).
 - **feat(payroll/multi-pays): plan comptable par pays + écritures salariales équilibrées (Closes #5256).** Nouveau registre `PayrollCountryChartOfAccounts` : les **21 pays de `CountryDefaults`** ont un plan comptable pour les écritures de paie (explicites DZ/MA/TN/SN/CI/CM/TR/FR/GB/US/CA — PCG/PCN, SYSCOHADA, Tekdüzen Hesap Planı, pratiques UK/US — + dérivés CEMAC→CM, CEDEAO→SN), chaque entrée portant `confidenceLevel` (production/pilot) et sa référence de référentiel. `PayrollAccountingExportService::journalLines()` (lecture seule, bulletins `validated` règle #2223) : écritures **équilibrées par construction** — D 641 (brut), D 645 (patronales), C 421 (net), C 431 (salariales + patronales), C 4421 (impôt + taxe forfaitaire via `flatPayrollTaxLabel`), C 425 (résidu avances/retenues) ; décomposition déterministe des déductions, débit OU crédit exclusif, traçabilité complète (`date`, `company_id`, `payroll_run_id`, `pay_slip_id`, `employee_id`, `reference` PAYROLL-RUN-{id}) — consommé par la Phase C #5239. `docs/payroll/MULTI_PAYS_PLAN_COMPTABLE.md` (tableau par pays + sources). 14 tests `PayrollAccountingExportJournalTest` (golden par pays calculés à la main, équilibre par bulletin/run, exclusion non validés, régularisation négative, TRIMF SN → 4421, pays dérivés, couverture 21/21 du registre).
 - **test(payroll/DZ): complément matrice #5244 — maladie + prime exonérée (Refs #5244).** `GoldenDzMatrix5244Test` : +7 cas golden — `computeSickLeaveAllowance` (carence 3 j, IJ 50 % J1-15 puis 100 % J16+, plafond 180 j — DZ_COMPLIANCE §8.2) et prime exonérée `is_taxable=false` via `computeNetBreakdown` 4e paramètre (§8.3).
