+++ api/app/Console/Commands/AutoCloseGeoSessionsCommand.php
@@ -4,22 +4,17 @@ declare(strict_types=1);
 
 namespace App\Console\Commands;
 
-use App\Core\Tenant\Domain\Models\Company;
-use App\Core\Tenant\TenantManager;
-use App\Modules\SmartAttendance\Domain\Models\GeoAttendanceSession;
 use Illuminate\Console\Command;
-use Illuminate\Support\Carbon;
 
 /**
- * Ferme automatiquement les sessions GPS qui sont restées ouvertes
- * trop longtemps (exemple : app tuée, perte réseau, crash).
+ * @deprecated ADR-0016 Phase 4 (#5355) — fusionnée dans `attendance:auto-close`
+ *             (une seule fermeture automatique). Cette commande est conservée
+ *             comme alias délégué pendant la transition ; elle sera supprimée
+ *             en Phase 5 (#5356).
  *
- * Multi-tenant : la commande itère TOUTES les companies (chaque tenant dans
- * son schéma via TenantManager::withinTenant) — sans itération, le scheduler
- * ne fermait que les sessions du schéma par défaut (#4797).
- *
- * À ajouter dans le scheduler Laravel :
- *   Schedule::command('smart-attendance:auto-close')->hourly();
+ * Ferme automatiquement les sessions GPS restées ouvertes trop longtemps
+ * (exemple : app tuée, perte réseau, crash). Délègue à
+ * `attendance:auto-close --sessions-only` (même cycle, même logique).
  */
 class AutoCloseGeoSessionsCommand extends Command
 {
@@ -28,104 +23,24 @@ class AutoCloseGeoSessionsCommand extends Command
                            {--company= : ID de la société (tenant) cible — sinon tous les tenants actifs}
                            {--dry-run  : Afficher sans fermer}';
 
-    protected $description = 'Ferme automatiquement les sessions GPS restées ouvertes (app tuée, crash réseau) — itère tous les tenants.';
+    protected $description = '@deprecated — alias de attendance:auto-close --sessions-only (ADR-0016 Phase 4, #5355). Ferme les sessions GPS restées ouvertes.';
 
-    public function handle(TenantManager $tenantManager): int
+    public function handle(): int
     {
-        $maxHours = (int) $this->option('hours');
-        $dryRun   = (bool) $this->option('dry-run');
-        $cutoff   = Carbon::now()->subHours($maxHours);
+        $args = ['--sessions-only' => true];
 
-        $companies = Company::query()
-            ->where('status', 'active')
-            ->when($this->option('company'), fn ($q, $id) => $q->whereKey($id))
-            ->orderBy('id')
-            ->get();
-
-        if ($companies->isEmpty()) {
-            $this->warn('Aucune société active — rien à fermer.');
-
-            return self::SUCCESS;
+        if ($this->option('hours') !== null && (int) $this->option('hours') !== 14) {
+            $args['--hours'] = (string) $this->option('hours');
         }
 
-        $totalClosed = 0;
-        $totalSkipped = 0;
-
-        foreach ($companies as $company) {
-            [$closed, $skipped] = $tenantManager->withinTenant(
-                $company,
-                fn (): array => $this->closeSessionsForTenant($company, $cutoff, $dryRun),
-            );
-
-            $totalClosed += $closed;
-            $totalSkipped += $skipped;
+        if ($this->option('company') !== null) {
+            $args['--company'] = (string) $this->option('company');
         }
 
-        if ($totalClosed === 0 && $totalSkipped === 0) {
-            $this->info("Aucune session GPS à fermer ({$companies->count()} tenant(s) parcouru(s)).");
-        } else {
-            $this->info("{$companies->count()} tenant(s) parcouru(s) — {$totalClosed} session(s) fermée(s), {$totalSkipped} en dry-run.");
+        if ((bool) $this->option('dry-run')) {
+            $args['--dry-run'] = true;
         }
 
-        if ($dryRun) {
-            $this->warn('[dry-run] Aucune modification effectuée.');
-        }
-
-        return self::SUCCESS;
-    }
-
-    /**
-     * Ferme les sessions GPS expirées dans le schéma du tenant courant.
-     *
-     * @return array{0: int, 1: int} [fermées, ignorées (dry-run)]
-     */
-    private function closeSessionsForTenant(Company $company, Carbon $cutoff, bool $dryRun): array
-    {
-        $sessions = GeoAttendanceSession::query()
-            ->whereNull('ended_at')
-            ->whereIn('status', [
-                GeoAttendanceSession::STATUS_DETECTED,
-                GeoAttendanceSession::STATUS_PENDING_VALIDATION,
-            ])
-            ->where('started_at', '<', $cutoff)
-            ->get();
-
-        if ($sessions->isEmpty()) {
-            return [0, 0];
-        }
-
-        $this->line(sprintf(
-            '  [%s] Sessions GPS à fermer : %d',
-            $company->slug,
-            $sessions->count(),
-        ));
-
-        $handled = 0;
-
-        foreach ($sessions as $session) {
-            $endedAt         = Carbon::now();
-            $durationSeconds = (int) $session->started_at->diffInSeconds($endedAt);
-
-            $this->line(sprintf(
-                '    → Session #%d / Employee #%d / démarrée à %s (durée: %dh)',
-                $session->id,
-                $session->employee_id,
-                $session->started_at->toDateTimeString(),
-                intdiv($durationSeconds, 3600),
-            ));
-
-            if (! $dryRun) {
-                $session->update([
-                    'ended_at'         => $endedAt,
-                    'duration_seconds' => $durationSeconds,
-                    'status'           => GeoAttendanceSession::STATUS_PENDING_VALIDATION,
-                    'validation_note'  => 'Fermée automatiquement (timeout système).',
-                ]);
-            }
-
-            $handled++;
-        }
-
-        return $dryRun ? [0, $handled] : [$handled, 0];
+        return $this->call('attendance:auto-close', $args);
     }
 }
