+++ api/app/Modules/Attendance/Application/Actions/RejectGeoSession.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace App\Modules\SmartAttendance\Application\Actions;
+namespace App\Modules\Attendance\Application\Actions;
 
 use App\Core\Auth\Domain\Models\Employee;
-use App\Modules\SmartAttendance\Domain\Models\GeoAttendanceSession;
+use App\Modules\Attendance\Domain\Models\GeoAttendanceSession;
 
 /**
  * Cas d'usage : rejet d'une session GPS par un manager ou RH.
