+++ api/app/Modules/Attendance/Application/DTOs/AttendanceModeConfigDTO.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Modules\SmartAttendance\Application\DTOs;
+namespace App\Modules\Attendance\Application\DTOs;
 
 final class AttendanceModeConfigDTO
 {
