+++ api/tests/Feature/Attendance/GeoMultiTenantIsolationTest.php
@@ -7,7 +7,7 @@ namespace Tests\Feature\SmartAttendance;
 use App\Core\Auth\Domain\Models\Employee;
 use App\Core\Tenant\Domain\Models\Company;
 use App\Modules\Planning\Domain\Models\Schedule;
-use App\Modules\SmartAttendance\Domain\Models\GeoAttendanceSession;
+use App\Modules\Attendance\Domain\Models\GeoAttendanceSession;
 use Illuminate\Support\Facades\Hash;
 use Laravel\Sanctum\Sanctum;
 use Tests\RefreshTenantDatabase;
@@ -19,7 +19,7 @@ use Tests\TestCase;
  * Vérifie qu'un employé/manager ne peut voir que les données
  * de sa propre company_id, jamais celles d'autres tenants.
  */
-class MultiTenantIsolationTest extends TestCase
+class GeoMultiTenantIsolationTest extends TestCase
 {
     use RefreshTenantDatabase;
