+++ api/tests/Feature/Attendance/GeofenceZoneServiceTest.php
@@ -8,7 +8,7 @@ use App\Core\Auth\Domain\Models\Employee;
 use App\Core\Tenant\Domain\Models\Company;
 use App\Core\Tenant\Domain\Models\Site;
 use App\Modules\Attendance\Infrastructure\Services\GeofenceZoneService;
-use App\Modules\SmartAttendance\Domain\Exceptions\OutsideGeofenceException;
+use App\Modules\Attendance\Domain\Exceptions\OutsideGeofenceException;
 use Illuminate\Support\Facades\Hash;
 use Tests\RefreshTenantDatabase;
 use Tests\TestCase;
